Move code from stream to device
